rename getter test file add more test in getters add more assertions in getter parts
